requirements.txt
imagehash
faiss-cpu
numpy
Pillow
transformers
torch
opencv-python
streamlit
matplotlib
